refactor: rename recruits -> recruit
